Remove ImageUtils and bring functionalities directly to adapter.
